Adds more domain events

+++ backend/auth/src/main/java/pt/ulisboa/tecnico/socialsoftware/auth/subscriptions/AuthUserUserSubscriptions.java
@@ -8,14 +8,16 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import pt.ulisboa.tecnico.socialsoftware.auth.domain.AuthUser;
+import pt.ulisboa.tecnico.socialsoftware.auth.domain.AuthUserState;
 import pt.ulisboa.tecnico.socialsoftware.auth.repository.AuthUserRepository;
 import pt.ulisboa.tecnico.socialsoftware.common.events.DeleteAuthUserEvent;
+import pt.ulisboa.tecnico.socialsoftware.common.events.RemoveUserFromTecnicoCourseExecutionEvent;
 
 import static pt.ulisboa.tecnico.socialsoftware.common.events.EventAggregateTypes.USER_AGGREGATE_TYPE;
 
 @Component
-public class UserSubscriptions {
-    private static final Logger logger = LoggerFactory.getLogger(UserSubscriptions.class);
+public class AuthUserUserSubscriptions {
+    private static final Logger logger = LoggerFactory.getLogger(AuthUserUserSubscriptions.class);
 
     @Autowired
     private AuthUserRepository authUserRepository;
@@ -24,7 +26,17 @@ public class UserSubscriptions {
         return DomainEventHandlersBuilder
                 .forAggregateType(USER_AGGREGATE_TYPE)
                 .onEvent(DeleteAuthUserEvent.class, this::deleteAuthUser)
+                .onEvent(RemoveUserFromTecnicoCourseExecutionEvent.class, this::removeUserFromCourseExecution)
                 .build();
+    }
+
+    public void removeUserFromCourseExecution(DomainEventEnvelope<RemoveUserFromTecnicoCourseExecutionEvent> event) {
+        logger.info("Received removeUserFromCourseExecution event!");
+        RemoveUserFromTecnicoCourseExecutionEvent removeUserFromTecnicoCourseExecutionEvent = event.getEvent();
+        AuthUser authUser = authUserRepository.findAuthUserById(removeUserFromTecnicoCourseExecutionEvent.getUserId())
+                .get();
+
+        authUser.getUserCourseExecutions().remove(removeUserFromTecnicoCourseExecutionEvent.getCourseExecutionId());
     }
 
     public void deleteAuthUser(DomainEventEnvelope<DeleteAuthUserEvent> event) {
@@ -36,9 +48,13 @@ public class UserSubscriptions {
                 .orElse(null);
 
         if (authUser != null) {
-            logger.info(authUser + "");
-            authUser.remove();
-            authUserRepository.delete(authUser);
+            if (authUser.getState().equals(AuthUserState.APPROVED)) {
+                authUser.remove();
+                authUserRepository.delete(authUser);
+            }
+            else {
+                authUser.setState(AuthUserState.REJECTED);
+            }
         }
     }
 }
